feat: split COLOR to theme
- theme에는 theme과 관련된 값만 있는게 적절하다 생각되어 COLOR는 따로 분리 함
